move the test to index.test.js Relates #21
